fix(eslint-errors): theme
