+++ packages/theme/src/create-memo-class.ts
@@ -1,14 +1,22 @@
 import { useMemo } from 'react';
 
-export function createMemoClass(func: (props: any) => string) {
-  return function useMemoClass(args?: any) {
-    const dependencies =
+export interface CreateMemoClassProps {
+  variant?: string;
+  size?: 'sm' | 'md' | 'lg' | string;
+  disabled?: boolean;
+  [key: string]: unknown;
+}
+
+export function createMemoClass(func: (props: CreateMemoClassProps) => string) {
+  return function useMemoClass(args: CreateMemoClassProps) {
+    const dependencies: unknown[] =
       typeof args === 'object' && args !== null
         ? Object.keys(args)
             .filter((key) => key !== 'theme')
-            .map((key) => args[key])
+            .map((key) => args[key as keyof CreateMemoClassProps])
         : [];
 
+    // eslint-disable-next-line react-hooks/exhaustive-deps
     return useMemo(() => func(args), dependencies);
   };
 }
