+++ packages/theme/src/index.ts
@@ -12,6 +12,9 @@ export * from './types';
 export type * from './types';
 export * from './units';
 export type * from './units';
+export * from './utils';
+export type * from './utils';
 export * from './fonts';
 export type * from './fonts';
 export * from './use-theme-queries';
+export * from './use-gui';
